Use derivative to evaluate basis functions at 0
This is because you get a 0/0 indeterminate form at the origin. But the
limit is still well-defined and we can evaluate it via the derivative.

Also, it turns out there was a bug in how I was assigning boundaries to
elements -- they were always excluded. Now, since we give zero special
treatment anyway, we consider elements to be (a, b], which simplifies
the code.